failure prevention work
Currently, if it's started without being connected to the internet, and
it failed at some point during the last upload and cached queries, it
will go into an infinite loop trying to send the queries, and them being
dumped back into the cache it's trying to flush.  This isn't a complete
fix yet.